+++ src/musiscape/__init__.py
@@ -8,4 +8,4 @@ itself. Sibling of `ambiscape <https://github.com/fourMs/ambiscape>`_
 """
 from .io import open_collection  # noqa: F401
 
-__version__ = "0.3.3"
+__version__ = "0.4.0"